+++ src/lib/transformers.js
@@ -4,7 +4,7 @@ import { PromiseEach } from 'utils/promise-each'
 import { PromiseMap } from 'utils/promise-map'
 /**
  * @typedef {Function} Validator
- * @desc Synchronous function that validates that given value is of the expected kind. Throws a {@link Schema~ValidationError} when not.
+ * @desc function (async permitted) that validates that given value is of the expected kind. Throws a {@link Schema~ValidationError} when not.
  * @param {*} value - The value being validated
  * @param {Object} [options]
  * @param {*} [options.state] - The state passed via the parse function
@@ -15,7 +15,7 @@ import { PromiseMap } from 'utils/promise-map'
 
 /**
  * @typedef {Function} Parser
- * @desc Synchronous function that performs custom logic possibly customized via settings that could transform given
+ * @desc function (async permitted) that performs custom logic possibly customized via settings that could transform given
  * value, throwing a {Schema~ValidationError} when error.
  * @param {*} value - The value being validated
  * @param {Object} [options]
@@ -27,7 +27,7 @@ import { PromiseMap } from 'utils/promise-map'
 
 /**
  * @typedef {Function} Caster
- * @desc Synchronous function that performs some logic attempting to cast given value into expected one. Returns the
+ * @desc function (async permitted) that performs some logic attempting to cast given value into expected one. Returns the
  * original value in case it could not be guessed.
  * @param {*} value - The value being casted
  * @param {Object} [options]